🎨 Formats files with prettier
